move validate to utils folder
