+++ cdm_reader_mapper/mdf_reader/utils/filereader.py
@@ -13,9 +13,9 @@ import xarray as xr
 
 from .. import properties
 from ..schemas import schemas
-from ..validate import validate
 from .configurator import Configurator
 from .utilities import convert_entries, decode_entries, validate_path
+from .validator import validate
 
 
 class FileReader:
