+++ cdm_reader_mapper/mdf_reader/utils/validator.py
@@ -9,10 +9,10 @@ import pandas as pd
 
 from cdm_reader_mapper.common.json_dict import get_table_keys
 
-from . import properties
-from .codes import codes
-from .schemas import schemas
-from .utils.utilities import convert_str_boolean
+from .. import properties
+from ..codes import codes
+from ..schemas import schemas
+from .utilities import convert_str_boolean
 
 
 def validate_datetime(elements, data):
